+++ src/BusinessLogic/Resources/js/RegisterModalController.js
@@ -10,9 +10,7 @@ if (!window.Packlink) {
      * @typedef {{
      *  name: string,
      *  code: string,
-     *  postal_code: string,
-     *  registration_link: string,
-     *  platform_country: string
+     *  postal_code: string
      * }} Country
      */
 
@@ -73,7 +71,7 @@ if (!window.Packlink) {
          */
         const handleCountrySelected = (country) => {
             modal.close();
-            Packlink.state.goToState('register', {country: country.code, platform_country: country.platform_country});
+            Packlink.state.goToState('register', {country: country.code});
         };
 
         /**