+++ CHANGELOG.md
@@ -3,6 +3,35 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## [3.2.0](https://github.com/packlink-dev/ecommerce_module_core/compare/v3.1.1...v3.2.0) - 2021-04-29
+**BREAKING CHANGES**
+Whitelabel project changes:
+
+### Added
+- Added BrandConfigurationService. Integration should register PacklinkConfigurationService instance for Packlink brand or adequate implementation for other brand(s) during the bootstraping.
+- Added PacklinkConfigurationService.
+- Added FileResolverService. Integration should initialize FileResolverService with an array of folders where source files should be searched for.
+- Added new method getLabels to \Packlink\BusinessLogic\CountryLabels\CountryService.
+
+### Changed
+
+- \Packlink\BusinessLogic\Language\TranslationService renamed to CountryService.
+- TranslationService::translate renamed to CountryService::getText.
+- Files from Resources/lang moved to Resources/countries. In integration, change path to translations to fit new folder names in core.
+- Configuration::getCurrentLanguage and Configuration::setCurrentLanguage changed to Configuration::getUICountryCode and Configuration::setUICountryCode.
+
+Following changes will work properly once BrandConfigurationService is registered in Bootstrap:
+- Removed hardcoded source value from ShippingServiceSearch.
+- Removed hardcoded platform code from Proxy, Draft and RegistrationRequest.
+- Removed platform country code from RegisterModalController.js.
+- Added platform_country to RegistrationController:getRegisterData response.
+- RegisterController.js function populateInitialValues() populates platform_country.
+- Registration link and platform country removed from CountryService::$supportedCountries.
+- Removed Packlink\BusinessLogic\Country\RegistrationCountry.
+- Added validation for platform country in RegistrationRequest::doValidate and UserAccountService::initializeUser.
+- Removed $supportedCountries from CountryService and WarehouseCountryService.
+- Packlink\BusinessLogic\Language\TranslationService renamed to CountryService and moved to CountryLabels directory.
+
 ## [3.1.1](https://github.com/packlink-dev/ecommerce_module_core/compare/v3.1.0...v3.1.1) - 2021-03-26
 ### Added
 - Added additional ISO codes to the postal code transformer.